requirements.txt
deeplake
llama-index
llama-index-readers-github
llama-index-vector-stores-deeplake
llama-index-embeddings-cohere
llama-index-llms-groq
python-dotenv